Brings in the mobile linked-probe work merged as #30, so this branch is current
before review.

One conflict, in .github/workflows/ci.yml, and both sides were wanted rather than
either winning: main added `timeout-minutes` to every job (so a wedged Qt mirror stops
burning toward the 360-minute default), while this branch replaced the python job with
a matrix covering every advertised interpreter plus the two MCP revisions. Resolved by
keeping the matrix and putting the timeout where the other jobs carry it.

Verified after resolution: all six jobs parse, the python job keeps its 7 legs and its
continue-on-error for the pre-release mcp-next path, static-probe keeps its 3, and
tests/test_ci_matrix_consistency.py -- which parses this very file -- still agrees with
the shipped BUILD_MATRIX.

C++ 20/20 on Qt 6.11.1. Python 266 passed; the two test_status_resource failures are
local only, from an unrelated process holding port 9222 on this machine.

+++ src/probe/accessibility/console_message_capture.cpp
@@ -72,7 +72,9 @@ QList<ConsoleMessage> ConsoleMessageCapture::messages(const QString& pattern, bo
   QList<ConsoleMessage> result;
 
   // Iterate in reverse order (newest first)
-  for (int i = m_messages.size() - 1; i >= 0; --i) {
+  // Narrow the bound once rather than the index: QList::at takes int on Qt5,
+  // so a qsizetype index would narrow at every call instead.
+  for (int i = static_cast<int>(m_messages.size()) - 1; i >= 0; --i) {
     const auto& cm = m_messages.at(i);
 
     // Filter by error level
